added image combining and image processing. can data from photo (somewhat lol)
can combine two transparent images so that it blends well. processes images to remove noise and try to isolate only the text. added subimage search with a given scale. makes the runtime a lot faster for particularly the stats window search because the scale for our skill window should be the same.